requirements.txt
tensorflow==2.10.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
prophet==1.1.4
influxdb-client==1.36.1
mysql-connector-python==8.0.33
schedule==1.2.0
requests==2.28.2
python-dotenv==1.0.0